requirements.txt
asgiref==3.7.2
click==8.1.7
Django==5.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
packaging==23.2
psycopg==3.1.18
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn==0.24.0.post1
